refactor: move component to index files
