--- wondrous-app/components/OnboardingDao/InviteCommunity/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './component';
