+++ wondrous-app/components/OnboardingDao/index.tsx
@@ -3,5 +3,5 @@ export { default as CreateDao } from './CreateDao';
 export { default as DaoCategory } from './DaoCategory';
 export { default as ImportTasks } from './ImportTasks';
 export { default as InviteCommunity } from './InviteCommunity';
-export { default as StepWrapper } from './StepWrapper';
 export { default as Review } from './Review';
+export { default as StepWrapper } from './StepWrapper';
